SKOTE-15: create DisplayNameByLocale trait for display name
