+++ app/app/Models/Category.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use App\Models\Traits\AdminTimestamp;
+use App\Models\Traits\DisplayNameByLocale;
 use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
@@ -11,7 +12,7 @@ use Illuminate\Support\Facades\Storage;
 
 class Category extends Model
 {
-    use HasFactory, SoftDeletes, AdminTimestamp;
+    use HasFactory, SoftDeletes, AdminTimestamp, DisplayNameByLocale;
 
     /**
      * The attributes that are mass assignable.
@@ -24,20 +25,6 @@ class Category extends Model
     ];
 
     /**
-     * Get & set categories.slug_name
-     *
-     * @return Attribute
-     */
-    public function slugName(): Attribute
-    {
-        return Attribute::make(
-            set: function (string $val) {
-                return slugifyModel($val, $this);
-            }
-        );
-    }
-
-    /**
      * Get full path for categories.image_file_name
      *
      * @return Attribute
@@ -46,20 +33,6 @@ class Category extends Model
     {
         return Attribute::make(
             get: fn() => $this->image_file_name ? Storage::url($this->image_file_name) : null
-        );
-    }
-
-    /**
-     * Get valid name by header locale
-     *
-     * @return Attribute
-     */
-    public function name(): Attribute
-    {
-        return Attribute::make(
-            get: function () {
-                return displayNameByLocale($this);
-            }
         );
     }
 }
