+++ app/app/Models/Product.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use App\Models\Traits\AdminTimestamp;
+use App\Models\Traits\DisplayNameByLocale;
 use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
@@ -12,7 +13,7 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Product extends Model
 {
-    use HasFactory, SoftDeletes, AdminTimestamp;
+    use HasFactory, SoftDeletes, AdminTimestamp, DisplayNameByLocale;
 
     /**
      * The attributes that are mass assignable.
@@ -25,20 +26,6 @@ class Product extends Model
     ];
 
     /**
-     * Get & set products.slug_name
-     *
-     * @return Attribute
-     */
-    public function slugName(): Attribute
-    {
-        return Attribute::make(
-            set: function (string $val) {
-                return slugifyModel($val, $this);
-            }
-        );
-    }
-
-    /**
      * Get full path for products.image_file_name
      *
      * @return Attribute
@@ -47,20 +34,6 @@ class Product extends Model
     {
         return Attribute::make(
             get: fn() => getFullPathToImage($this, 'image_file_name')
-        );
-    }
-
-    /**
-     * Get valid name by header locale
-     *
-     * @return Attribute
-     */
-    public function name(): Attribute
-    {
-        return Attribute::make(
-            get: function () {
-                return displayNameByLocale($this);
-            }
         );
     }
